adjust precision

+++ overhead-test/src/Results/CounterResults.cs
@@ -5,9 +5,9 @@ namespace SignalFx.OverheadTest.Results;
 // MB is preferred.
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 internal record CounterResults(
-    float AverageTotalCpuPercentage,
-    float AverageWorkingSetMB,
-    float AverageTimeInGcPercentage,
-    float MinHeapUsedMB,
-    float MaxHeapUsedMB,
-    float TotalAllocatedMB);
+    double AverageTotalCpuPercentage,
+    double AverageWorkingSetMB,
+    double AverageTimeInGcPercentage,
+    double MinHeapUsedMB,
+    double MaxHeapUsedMB,
+    double TotalAllocatedMB);
