+++ overhead-test/src/Results/K6Results.cs
@@ -1,7 +1,7 @@
 ﻿namespace SignalFx.OverheadTest.Results;
 
 internal record K6Results(
-    float IterationDurationAvg,
-    float IterationDurationP95,
-    float RequestDurationAvg,
-    float RequestDurationP95);
+    double IterationDurationAvg,
+    double IterationDurationP95,
+    double RequestDurationAvg,
+    double RequestDurationP95);
